requirements.txt
appium-python-client
boto3
tiktoken
fastapi
pydantic
uvicorn
PyGithub
GitPython
pyyaml